move to /data/remote/body directory
